cast to types expected by Win32 API
```
src/agent.c:314:..: warning C4267: '=': conversion from 'size_t' to 'DWORD', possible loss of data
src/agent.c:314:18: warning: conversion from 'size_t' {aka 'long long unsigned int'} to 'DWORD' {aka 'long unsigned int'} may change value [-Wconversion]
src/agent.c:314:20: warning: implicit conversion loses integer precision: 'unsigned long long' to 'DWORD' (aka 'unsigned long') [-Wshorten-64-to-32]
src/wincng.c:671:..: warning C4267: 'function': conversion from 'size_t' to 'ULONG', possible loss of data
src/wincng.c:671:39: warning: conversion from 'size_t' {aka 'long long unsigned int'} to 'ULONG' {aka 'long unsigned int'} may change value [-Wconversion]
src/wincng.c:671:39: warning: implicit conversion loses integer precision: 'size_t' (aka 'unsigned long long') to 'ULONG' (aka 'unsigned long') [-Wshorten-64-to-32]
src/wincng.c:2552:33: warning: conversion from 'long long unsigned int' to 'DWORD' {aka 'long unsigned int'} may change value [-Warith-conversion]
```

+++ src/agent.c
@@ -311,7 +311,7 @@ agent_transact_pageant(LIBSSH2_AGENT *agent, agent_transaction_ctx_t transctx)
                        transctx->request_len);
 
     cds.dwData = PAGEANT_COPYDATA_ID;
-    cds.cbData = 1 + strlen(mapname);
+    cds.cbData = (DWORD)(1 + strlen(mapname));
     cds.lpData = mapname;
 
     id = SendMessage(hwnd, WM_COPYDATA, (WPARAM) NULL, (LPARAM) &cds);
